SCRD-2781 Document installation and repo prep
In order to make the Cloud installer and updates media available to the
all nodes in the deployment beyond just the deployer node, certain
manual steps must be taken. This patch updates the GUI installer guide
to instruct users to take these extra preparation steps. These steps are
very similar to steps 3, 4, and 5 of setting up the administration
server for Cloud 7[1].

[1] http://docserv.nue.suse.com/documents/Cloud7/suse-openstack-cloud-deployment/html/part.depl.admserv.html