Several changes to security guide:
Corrected typo in Barbican documentation.
Changed "hos" to "ardana" in several paths to files and file names as appropriate.
Changed "hlm" to "ardana" or removed it as appropriate.
Changed "HPE" example names in certificates to "SUSE".
Updated link to ESKM to new MF link instead of the old HPE link (which forwards to the new MF link today but that will not last).
Clarified that SELinux is only available for the Red Hat compute node.
Added "openstack role add --user admin --project admin cinder_admin" to instructions to configure DAR encryption so they still work with the updated policy.json files.
Updated expiry shown for the default cert to "notAfter=Feb 12 01:18:46 2019 GMT".