Fix the swift container-sync documentation
The realm keys are used by container-sync to create a trust
with another cluster in the realm. They _should_ never be used
as the keys used in the container metadata when setting up the
sync between containers.

Container sync will use both the realm key(s) and the key specified
in the container metadata to generate a HMAC to authenticate the
container-sync request to push objects to the destination container.

The doc specifically said to use the realm key when setting up
sync between containers, which is not only incorrect but insecure.
General users should not know the realm keys. Ops should set up the
trust (realm keys). Users can use /info to see other clusters
and will generate their own keys to set up container sync relationships
with containers.