Remove references to diskimage-builder
In general, the default ramdisk image that we provide as part of the
product should be sufficient and customers should not need to build
their own ramdisk, nor should they have to create a special boot ISO
image because automatic boot ISO creation should work on its own. This
change removes the instructions to create a special image for Proliant
support since our ramdisk image has that built in, and removes
references to the manual boot ISO creation. Additionally clean up other
unnecessary references to DIB.

In the future, we should come up with a cohesive story around allowing
users to customize their own images using either Kiwi or
diskimage-builder and a freely distributed base image in OBS.