Add references to DIB back to guide
This partially reverts #761. It's not forbidden to mention
diskimage-builder or other freely licensed distros, so long as we call
out that special customizations to ramdisk images or boot ISOs can't be
supported. This also corrects an error that #761 had removed, which
called out 'suse' as a DIB element. 'suse' is not an element available
in DIB (it's 'opensuse'), and moreover the section called for a special
'-signed' element which is not available for opensuse, so we need to
reference the 'ubuntu-signed' element.