Use patches instead of sed
In order to generate documentation that cleanly removes all references
to installation and contributor guides, including those nested inside of
paragraphs of text, requires generating complicated and hard-to-maintain
sed scripts.  Replace that logic with the more straightforward approach
of carrying patches to each of the projects and applying them at doc
build time to reach the desired state. Further updates to these
documents becomes a matter of updating the patch file rather than
writing code to elicit the same result.