Adding skeleton for new CLM POC Guide
This guide is not even close to being ready. But this is the skeleton
that is available for public review. This work is being collaborated
on with Ron Terry from the Training team. He has provided a workbook
on deploying SUSE OpenStack Cloud with Cloud Lifecycle Manager that
will provide the basis for this content.

The structure that Ron has implemented was the work we did and discussed
at the Doc Day at SUSECON in April 2019.