Supplement: Fix guide grouping on documentation.suse.com (trivial)
On documentation.suse.com/soc/9, the Supplement Guide is currently shown
on  two lines like this:

Supplement Guide   HTML
Supplement Guide   Single-HTML PDF EPUB

This commit aligns the profiling values of DC-...clm-all and
DC-...supplement. This has the effect of fixing the grouping, so
everything is in one line on the web page.