Update incorrect Manila install/setup instructions (SOC-10975)
Per https://bugzilla.suse.com/show_bug.cgi?id=1149860 (also https://jira.suse.com/browse/SOC-10975 ), the Cloud 8 documentation for Manila is wrong. Using the Cloud 9 documentation (which is correct) is the present workaround. This change copies the Cloud 9 Manila docs into Cloud 8. The only change from the Cloud 9 content is a replacement of &product; with &productname; as the former does not exist in Cloud 8.